fix: extrait les enums exportés dans le package type
