+++ lib/outputs/history.js
@@ -1,4 +1,4 @@
-const { CertificationState } = require('./record')
+const { CertificationState, EventType } = require('../enums')
 
 /**
  * @typedef {import('../providers/types/cartobio').CartoBioUser} CartoBioUser
@@ -7,19 +7,6 @@ const { CertificationState } = require('./record')
  * @typedef {import('./types/record').NormalizedRecord} NormalizedRecord
  * @typedef {import('./types/history').HistoryEntry} HistoryEntry
  */
-
-/**
- * @enum {String}
- */
-const EventType = {
-  CERTIFICATION_STATE_CHANGE: 'CertificationStateChange',
-  FEATURE_COLLECTION_CREATE: 'FeatureCollectionCreation',
-  FEATURE_COLLECTION_DELETE: 'FeatureCollectionDeletion',
-  FEATURE_COLLECTION_UPDATE: 'FeatureCollectionUpdate',
-  FEATURE_CREATE: 'FeatureCreation',
-  FEATURE_DELETE: 'FeatureDeletion',
-  FEATURE_UPDATE: 'FeatureUpdate'
-}
 
 function isAfterCertificationState (referenceState) {
   const stateValues = Object.values(CertificationState)
