+++ lib/outputs/history.test.js
@@ -1,5 +1,5 @@
-const { collectFeatureIdsFromPayload, createNewEvent, EventType } = require('./history.js')
-const { CertificationState } = require('./record')
+const { collectFeatureIdsFromPayload, createNewEvent } = require('./history.js')
+const { CertificationState, EventType } = require('../enums')
 
 const collection = {
   type: 'FeatureCollection',
