+++ lib/outputs/record.js
@@ -1,5 +1,6 @@
 const { featureCollection } = require('@turf/helpers')
 const { normalizeParcelle } = require('./features')
+const { EtatProduction } = require('../enums')
 
 /**
  * @typedef {import('../providers/types/agence-bio').AgenceBioOperator} AgenceBioOperator
@@ -9,31 +10,6 @@ const { normalizeParcelle } = require('./features')
  * @typedef {import('./types/record').NormalizedRecord} NormalizedRecord
  * @typedef {import('./types/record').NormalizedRecordSummary} NormalizedRecordSummary
  */
-
-/**
- * @readonly
- * @enum {String}
- */
-const CertificationState = {
-  OPERATOR_DRAFT: 'OPERATOR_DRAFT', // Phase 2
-  AUDITED: 'AUDITED', // Phase 3
-  PENDING_CERTIFICATION: 'PENDING_CERTIFICATION', // Phase 4
-  CERTIFIED: 'CERTIFIED' // Phase 5
-}
-
-/**
- * @readonly
- * @enum {String}
- */
-const EtatProduction = {
-  C0: 'CONV',
-  CONV: 'CONV',
-  NB: 'CONV',
-  C1: 'C1',
-  C2: 'C2',
-  C3: 'C3',
-  AB: 'AB'
-}
 
 /**
  * Normalize a record and add operator data
@@ -82,7 +58,5 @@ function normalizeEtatProduction (etat) {
 module.exports = {
   normalizeRecord,
   normalizeRecordSummary,
-  normalizeEtatProduction,
-  EtatProduction,
-  CertificationState
+  normalizeEtatProduction
 }
