+++ lib/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@agencebio/cartobio-types",
-  "version": "1.4.1",
+  "version": "1.4.2",
   "main": "types.d.ts",
   "repository": {
     "type": "git",
