+++ lib/providers/cartobio.js
@@ -11,13 +11,14 @@ const { SocksProxyAgent } = require('socks-proxy-agent')
 
 const pool = require('../db.js')
 const config = require('../config.js')
+const { EtatProduction, CertificationState, EventType } = require('../enums')
 const { parsePacDetailsFromComment, fetchOperatorByNumeroBio } = require('./agence-bio.js')
 const { populateWithRecords } = require('../outputs/operator.js')
-const { normalizeRecord, normalizeRecordSummary, EtatProduction, CertificationState, normalizeEtatProduction } = require('../outputs/record.js')
+const { normalizeRecord, normalizeRecordSummary, normalizeEtatProduction } = require('../outputs/record.js')
 const { randomUUID } = require('crypto')
 const { fromCodePacStrict } = require('@agencebio/rosetta-cultures')
 const { fromCepageCode } = require('@agencebio/rosetta-cultures/cepages')
-const { EventType, createNewEvent } = require('../outputs/history.js')
+const { createNewEvent } = require('../outputs/history.js')
 const { NotFoundApiError } = require('../errors.js')
 const { InvalidRequestApiError } = require('../errors.js')
 const { getRandomFeatureId, populateWithMultipleCultures } = require('../outputs/features.js')
