+++ lib/providers/telepac.js
@@ -6,7 +6,7 @@ const { XMLParser } = require('fast-xml-parser')
 const { getRandomFeatureId } = require('../outputs/features.js')
 const { obj } = require('through2')
 const { randomUUID } = require('crypto')
-const { EtatProduction } = require('../outputs/record.js')
+const { EtatProduction } = require('../enums')
 const { fromCodePacStrict } = require('@agencebio/rosetta-cultures')
 const { featureCollection, feature: Feature } = require('@turf/helpers')
 const intersect = require('@turf/intersect').default
