+++ lib/routes/records.js
@@ -1,4 +1,4 @@
-const { EtatProduction } = require('../outputs/record.js')
+const { EtatProduction } = require('../enums')
 
 const createRecordSchema = {
   schema: {
