+++ lib/types.d.ts
@@ -2,5 +2,4 @@ export type {HistoryEntry} from './outputs/types/history.d.ts';
 export type {NormalizedRecord, NormalizedRecordSummary} from './outputs/types/record.d.ts';
 export type {AgenceBioNormalizedOperator, AgenceBioNormalizedOperatorWithRecord} from './outputs/types/operator.d.ts';
 export type {CartoBioFeature, CartoBioFeatureCollection, CartoBioFeatureProperties} from './outputs/types/features.d.ts';
-export {EtatProduction, CertificationState} from './outputs/record.js';
-export {EventType} from './outputs/history.js';
+export {EtatProduction, CertificationState, EventType} from './enums.js';
